refactor: significantly reduce costs_conting blast radius
The goal here is to make sure that only the leaf crates (primitives-core
and vm-logic) and the topmost crate (params-estimator) know about this
feature.

Before this PR, every intermediary crate would have to know about this
feature as well.

To fix this, we encapsulate all conditional compilation logic inside
ProfileData type, while making sure it compiles to no-op unless the
feature is enabled.